chore/refactor-shared-utils: refactor: migrate utility functions to fileUtils directory and remove obsolete files
